feat: 기존 useEffect를 분리된 contents로 이동
