+++ src/pages/MeetingVote/MeetingVote.tsx
@@ -1,7 +1,7 @@
 import { UIEvent, useEffect, useRef, useState } from 'react';
 import { useResizeDetector } from 'react-resize-detector';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilValue } from 'recoil';
 
 import { ScrollDownFloatingButton } from '../../components/buttons/ScrollDownFloatingButton';
 import { Loading } from '../../components/Loading';
@@ -9,8 +9,6 @@ import { Page } from '../../components/pageLayout';
 import { useMeeting } from '../../hooks/useMeeting';
 import { useVotings } from '../../hooks/useVotings';
 import { currentUserStateFamily } from '../../stores/currentUser';
-import { currentUserVotingSlotsState } from '../../stores/currentUserVotingSlots';
-import { votingsState as votingRecoilState } from '../../stores/voting';
 import MeetingVoteContents from '../../templates/MeetingVote/MeetingVoteContents';
 import MeetingVoteFooter from '../../templates/MeetingVote/MeetingVoteFooter';
 import MeetingVoteHeader from '../../templates/MeetingVote/MeetingVoteHeader';
@@ -18,17 +16,15 @@ import MeetingVoteHeader from '../../templates/MeetingVote/MeetingVoteHeader';
 function MeetingVote() {
   const [searchParams] = useSearchParams();
   const pageElementRef = useRef<HTMLDivElement>(null);
-  const { meeting, meetingId, isFetching: isMeetingFetching } = useMeeting();
-  const { votings, isFetching: isVotingsFetcing } = useVotings();
+  const { meetingId, isFetching: isMeetingFetching } = useMeeting();
+  const { isFetching: isVotingsFetcing } = useVotings();
   const isFetching = isMeetingFetching && isVotingsFetcing;
 
   const currentUser = useRecoilValue(currentUserStateFamily(meetingId));
-  const setCurrentUserVotingSlots = useSetRecoilState(currentUserVotingSlotsState);
   const isNewUser = !currentUser;
 
   const [hasMoreBottomScroll, setHasMoreBottomScroll] = useState<boolean>(false);
 
-  const setVotingsState = useSetRecoilState(votingRecoilState);
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -37,15 +33,6 @@ function MeetingVote() {
       navigate(`/meetings/${meetingId}`);
     }
   }, [isNewUser, meetingId, navigate, searchParams]);
-
-  useEffect(() => {
-    if (meeting && votings) {
-      setVotingsState(votings);
-      const currentUserVoting = votings.find((voting) => voting.id === currentUser?.id);
-      const currentUserVotingSlots = currentUserVoting?.[meeting.type];
-      setCurrentUserVotingSlots(currentUserVotingSlots ?? []);
-    }
-  }, [meetingId, setVotingsState, setCurrentUserVotingSlots, currentUser, votings, meeting]);
 
   useResizeDetector({
     targetRef: pageElementRef,
@@ -79,10 +66,6 @@ function MeetingVote() {
 
   if (isFetching) {
     return <Loading />;
-  }
-
-  if (!meeting) {
-    return null;
   }
 
   return (
